Change built in pattern to an enum fro code to gast
