+++ js_built_in_functions.py
@@ -1,0 +1,5 @@
+import enum
+
+class js_built_in_functions(enum.Enum):
+    push = "appendStatement"
+    pop = "popStatement"
