+++ py_built_in_functions.py
@@ -1,0 +1,5 @@
+import enum
+
+class py_built_in_functions(enum.Enum):
+    append = "appendStatement"
+    pop = "popStatement"
